Use more fixed size ints

+++ components/esm3/loadcell.hpp
@@ -34,7 +34,7 @@ namespace ESM
         RefNum mRefNum;
 
         // Coordinates of target exterior cell
-        int mTarget[2];
+        int32_t mTarget[2];
 
         // The content file format does not support moving objects to an interior cell.
         // The save game format does support moving to interior cells, but uses a different mechanism
@@ -153,13 +153,13 @@ namespace ESM
             ESMReader& esm, bool saveContext = true); // Load everything, except NAME, DATAstruct and references
 
         void save(ESMWriter& esm, bool isDeleted = false) const;
-        void saveTempMarker(ESMWriter& esm, int tempCount) const;
+        void saveTempMarker(ESMWriter& esm, int32_t tempCount) const;
 
         bool isExterior() const { return !(mData.mFlags & Interior); }
 
-        int getGridX() const { return mData.mX; }
+        int32_t getGridX() const { return mData.mX; }
 
-        int getGridY() const { return mData.mY; }
+        int32_t getGridY() const { return mData.mY; }
 
         bool hasWater() const { return ((mData.mFlags & HasWater) != 0) || isExterior(); }
 
@@ -172,7 +172,7 @@ namespace ESM
         // somewhere other than the file system, you need to pre-open the
         // ESMReader, and the filename must match the stored filename
         // exactly.
-        void restore(ESMReader& esm, int iCtx) const;
+        void restore(ESMReader& esm, size_t iCtx) const;
 
         std::string getDescription() const;
         ///< Return a short string describing the cell (mostly used for debugging/logging purpose)
